Remove legacy metrics client from podautoscaler
